requirements.txt
flask
requests
openai==0.28
pillow
pandas
markdown
numpy
faker
pytesseract
scikit-learn

